requirements.txt
# Medical Image Segmentation - Requirements
# For local development (lightweight)

# Core
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0
opencv-python>=4.5.0
matplotlib>=3.4.0
scikit-learn>=0.24.0

# Dataset Download
kagglehub>=0.2.0

# Jupyter
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# For Colab training (install separately in Colab):
# torch>=1.10.0
# torchvision>=0.11.0
# timm>=0.5.0
# albumentations>=1.1.0
# tqdm>=4.62.0
